scripts/requirements.txt
asciidoc
pyyaml
Mako
jsonschema